Fix bug regarding recalculation of locations
Clicking on a region pill causes a recalculation in the chosen options.
This means if North is clicked the result may be different from when it
is then clicked again. This fixes that problem to only perform a rng
recalc on date change.

Next, the application was built so that a lower component DateNavigation
affected global state. In reality, we want that component to control the
date selections, and then propagate up to the higher App component to
do the necessary processing. With this, the App component has full
control over the selected date and can make timezone changes as needed.
Which in this case is always going to be set to America/Chicago for now.